+++ spec/spec_helper.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
 
+require "stringio"
+require "tempfile"
+
 require_relative "../lib/lumberjack_json_device"
 
 RSpec.configure do |config|
@@ -8,4 +11,19 @@ RSpec.configure do |config|
   config.default_formatter = "doc" if config.files_to_run.one?
   config.order = :random
   Kernel.srand config.seed
+end
+
+def silence_deprecations
+  save_warning = Warning[:deprecated]
+  save_verbose = $VERBOSE
+  begin
+    Warning[:deprecated] = false
+    $VERBOSE = false
+    begin
+      yield
+    ensure
+      Warning[:deprecated] = save_warning
+      $VERBOSE = save_verbose
+    end
+  end
 end